% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup.R
\name{sup}
\alias{sup}
\title{Convert letters to Unicode superscript}
\usage{
sup(letter)
}
\arguments{
\item{letter}{A character vector of single lowercase letters}
}
\value{
A character vector of superscript glyphs; errors if a letter has no
mapping
}
\description{
Looks up footnote-marker letters in \link{superscript_letters}.
}
\examples{
paste0("Year", sup("a"))
sup(c("a", "b"))
}
