% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minnowTables-package.R
\docType{package}
\name{minnowTables-package}
\alias{minnowTables}
\alias{minnowTables-package}
\title{minnowTables: R Package for Excel table formatting.}
\description{
Helps create excel tables in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Trinity-Minnow/minnowTables}
  \item Report bugs at \url{https://github.com/Trinity-Minnow/minnowTables/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
