% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shading_styles.R
\docType{data}
\name{shading_styles}
\alias{shading_styles}
\alias{blue_style}
\alias{green_style}
\title{Cell shading styles}
\format{
An openxlsx Style object

An object of class \code{Style} of length 1.

An object of class \code{Style} of length 1.
}
\usage{
blue_style

green_style
}
\description{
Fill styles reused for shading table cells: blue marks a
positive/increase, green a negative/decrease.
}
\examples{
\dontrun{
openxlsx::addStyle(wb, "Table 1", blue_style, rows = 5, cols = 3, stack = TRUE)
}
}
\keyword{datasets}
